Fix an off-by-one error.
HTTP ranges are inclusive. Previously, the range could go from 0 to the
length of the content. This commit fixes this; the range now maxes out
at the length of the content minus one.

Users should note that when using this library, they should always index
slices like b[r.Start : r.Stop + 1] as slice indexing is exclusive on
the right-hand-side. This is because I feel that the library should
preserve the range in the header as faithfully as possible while still
providing conveniences like bounds and merging.